Model_Pipeline/scripts/requirements.txt
transformers
datasets
faiss-cpu 
google-cloud-storage